update the schema and template
